requirements.txt
fastapi
uvicorn[standard]


pydantic
pydantic-settings
python-dotenv

alembic
sqlalchemy

psycopg2-binary

python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2